changed ISBN to bigint in UserBookStatus table

+++ backend/database/DDL.sql
@@ -48,7 +48,7 @@ CREATE TABLE Books (
 CREATE TABLE UserBookStatus (
     statusID int NOT NULL AUTO_INCREMENT,
     userID int NOT NULL,
-    ISBN int NOT NULL,
+    ISBN bigint NOT NULL,
     readingStatus ENUM('Completed', 'In Progress', 'Not Started', 'Did Not Finish'),
     startDate DATE,
     finishDate DATE,
